python/requirements.txt
numpy>=1.24.0
pyzmq>=25.1.0
onnxruntime>=1.15.0
torch>=2.0.0
requests>=2.31.0
pytest>=7.4.0
